requirements_bullseye.txt
pyzmq==17.1.2
configargparse==0.13.0
voluptuous==0.11.1
urwid==2.0.1
sqlalchemy==1.3.15
psycopg2==2.8.5
chameleon==3.6.2
python-dateutil==2.8.1
